Feature/adding jwt claims

+++ flask_application/section11/code/resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
 
@@ -32,7 +32,11 @@ class Item(Resource):
             return {'message': 'An error occurred inserting the item'}, 500  # internal server error
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
